+++ crmModule/templates/base.html
@@ -34,6 +34,7 @@
             <li>
                 <a href="/signup/">SignUp</a>
             </li>
+            {% endif %}
             <li>
                 <a href="/signin/">SignIn</a>
             </li> -->
