--- crmModule/templates/register_user.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous"></script>
-    <link rel="stylesheet" type="text/css" href="css/main.css"> 
-</head>
-
-    <body>
-
-{% block content %}
-
-<h1>Register user</h1>
-
-<p>{{error}}</p>
-
-<form action="/user/" method="POST">
-
-    {% csrf_token %}
-    {{form.as_p}}
-
-    <button class="btn btn-primary">
-        Registrarse
-    </button>
-</form>
-{% endblock %}
