+++ crmModule/urls.py
@@ -4,9 +4,9 @@ from .views import user_views, views, sponsor, event, getInformation
 
 urlpatterns = [
     path('home/', views.home, name='home'),
-    path('signup/', views.signup, name='signup'),
-    path('', views.signin,name='signin'),
-    path('logout/', views.signout,name='signout'),
+    path('signup/', user_views.signup, name='signup'),
+    path('', user_views.signin,name='signin'),
+    path('logout/', user_views.signout,name='signout'),
     path('menu/', views.menu,name='menu'),
     path('sponsor/', sponsor.sponsor, name='sponsor'),
     path('sponsor/register/', sponsor.register_sponsor,name='register_sponsor'),
@@ -22,7 +22,7 @@ urlpatterns = [
     path('event/<int:event_id>/delete', event.delete_event,name='delete_event'),
     path('event/l/', event.show_events, name='show_events'),
     #path('event/register/event/register/', event.register_event,name='register_event'),
-    path('user/',user_views.signup,name='user'),
+    #path('user/',user_views.signup,name='user'),
     #path('event/register/event/register/', event.register_event,name='register_event'),
 
 ]
