+++ crmModule/views/user_views.py
@@ -2,11 +2,13 @@ from django.shortcuts import render, redirect
 from ..forms import UserForm
 from ..models import Official
 from django.contrib.auth import login, logout, authenticate
+from django.contrib.auth.forms import AuthenticationForm
+
 
 def signup(request):
 
     if request.method == 'GET':
-        return render(request, 'register_user.html',
+        return render(request, 'signup.html',
                 {
                     'form': UserForm
                 })
@@ -19,15 +21,33 @@ def signup(request):
                 return redirect('signin')
             
             except:
-                return render(request, 'register_user.html',
+                return render(request, 'signup.html',
                               {
                                   'form': UserForm,
                                   "error": 'Username already exist'
                               })
         else:
-            return render(request, 'register_user.html',
+            return render(request, 'signup.html',
                               {
                                   'form': UserForm,
                                   "error": 'Password do not match'
                               })
+        
+def signin(request):
+     if request.method == 'GET':
+        return render(request, 'signin.html',
+                  {
+                      'form': AuthenticationForm
+                  })
+     else:
+        user= authenticate(request, username=request.POST['username'], password=request.POST['password'])
+        if user is None:
+            return render(request,'signin.html', {'form': AuthenticationForm, 'error': 'USUARIO O CONTRASEÑA INCORRECTA' })
+        else:
+            login(request, user)
+            return redirect('menu')
+        
+def signout(request):
+    logout(request)
+    return redirect('home')
